v1.5.0-beta.7
v1.5.0-beta.7

+++ diskover-bot-launcher.sh
@@ -9,28 +9,23 @@
 #
 
 ############# EDIT BELOW FOR YOUR ENVIRONMENT #############
-
 # paths and default config values below
-PYTHON=python # path to python command or python if in PATH
-DISKOVERBOT=./diskover_worker_bot.py # path to diskover_worker_bot.py
+
+# path to python command or python if in PATH
+PYTHON=python
+# path to diskover_worker_bot.py
+DISKOVERBOT=./diskover_worker_bot.py
 # number of bots to start (cores x 2 might be good start)
 WORKERBOTS=8
 # run bots in burst mode (quit when all jobs done)
 BURST=FALSE
-# run bots with more output
-VERBOSE=FALSE
-# run bots with less output
-QUIET=FALSE
-# bot log file directory and file prefix
-# logs will be named BOTLOG_timestamp
-BOTLOG=/tmp/diskover_bot_log
 # file to store bot pids
 BOTPIDS=/tmp/diskover_bot_pids
 
 ###########################################################
 
 
-VERSION="1.0"
+VERSION="1.1"
 
 function printhelp {
     echo "Usage: $(basename $0) [OPTION] [ROOTDIR]"
@@ -38,11 +33,9 @@ function printhelp {
     echo "Options:"
     echo
     echo "  -w n    number of worker bots to start (default $WORKERBOTS)"
-    echo "  -b      burst mode (bots quit when no more jobs"
-    echo "  -v      more output (verbose)"
-    echo "  -q      less output (quiet)"
-    echo "  -k      kill all bots"
+    echo "  -b      burst mode (bots quit when no more jobs)"
     echo "  -s      show all bots running"
+    echo "  -k      kill all bots"
     echo "  -V      displays version and exits"
     echo "  -h      displays this help message and exits"
     echo
@@ -54,13 +47,11 @@ function printhelp {
 
 KILLBOTS=FALSE
 SHOWBOTS=FALSE
-while getopts :h?w:bvqskV opt; do
+while getopts :h?w:bskV opt; do
     case "$opt" in
     h) printhelp; exit 0;;
     w) WORKERBOTS=$OPTARG;;
     b) BURST=TRUE;;
-    v) VERBOSE=TRUE;;
-    q) QUIET=TRUE;;
     s) SHOWBOTS=TRUE;;
     k) KILLBOTS=TRUE;;
     V) echo "$0 v$VERSION"; exit 0;;
@@ -77,7 +68,7 @@ function banner {
    |    \`   \\  |\\___ \\|    <  <_> )   /\\  ___/|  | \\/ (='.'=)
   /_______  /__/____  >__|_ \\____/ \\_/  \\___  >__|   (\"\)_(\"\)
           \\/        \\/     \\/               \\/
-                worker bot launcher v$VERSION
+                Worker Bot Launcher v$VERSION
                 https://github.com/shirosaidev/diskover
                 \"Crawling all your stuff, core melting time\"$(tput sgr 0)
 
@@ -85,19 +76,13 @@ function banner {
 }
 
 function startbots {
-	echo "starting $WORKERBOTS worker bots in background..."
+	echo "Starting $WORKERBOTS worker bots in background..."
     ARGS=""
     if [ $BURST == TRUE ]; then
-        ARGS+=" -b "
-    fi
-    if [ $VERBOSE == TRUE ]; then
-        ARGS+=" -v "
-    fi
-    if [ $QUIET == TRUE ]; then
-        ARGS+=" -q "
+        ARGS+="-b"
     fi
     for (( i = 1; i <= $WORKERBOTS; i++ )); do
-        $PYTHON $DISKOVERBOT $ARGS > ${BOTLOG}_${i}_$(date +%H:%M:%S) 2>&1 &
+        $PYTHON $DISKOVERBOT $ARGS > /dev/null 2>&1 &
         echo "$(hostname -s).$! (pid $!)"
         echo "$!" >> $BOTPIDS
     done
@@ -108,25 +93,33 @@ function startbots {
 banner
 
 if [ $KILLBOTS == TRUE ]; then
-    echo "killing worker bot pids:"
-    for i in `cat $BOTPIDS`; do
-        echo "$i"
-        kill $i
-    done
-    echo "all worker bots have been sent shutdown command"
-    rm $BOTPIDS
+    if [ -f $BOTPIDS ]; then
+        echo "Killing worker bot pids:"
+        for PID in `cat $BOTPIDS`; do
+            echo "$PID"
+            if `ps -p $PID > /dev/null 2>&1`; then
+                kill $PID
+            fi
+        done
+        echo "All worker bots have been sent shutdown command"
+        rm $BOTPIDS
+    else
+        echo "No worker bots running or pid file can't be found"
+    fi
 elif [ $SHOWBOTS == TRUE ]; then
     if [ -f $BOTPIDS ]; then
-        echo "running worker bots:"
-        for i in `cat $BOTPIDS`; do
-            echo "$(hostname -s).$i (pid $i)"
+        echo "Running worker bots:"
+        for PID in `cat $BOTPIDS`; do
+            if `ps -p $PID > /dev/null 2>&1`; then
+                echo "$(hostname -s).$PID (pid $PID)"
+            fi
         done
     else
-        echo "no worker bots running"
+        echo "No worker bots running"
     fi
 else
     startbots
-    echo "all worker bots have started, logging to $BOTLOG"
-    echo "worker pids have been stored in $BOTPIDS, use -k flag to shutdown workers"
-    echo "exiting, sayonara!"
+    echo "All worker bots have started"
+    echo "Worker pids have been stored in $BOTPIDS, use -k flag to shutdown workers"
+    echo "Exiting, sayonara!"
 fi
